requirements.txt
pipecat-ai[openai,deepgram]
deepgram-sdk
daily-python
python-dotenv
aiohttp
loguru
pillow
soundfile
resampy
